# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dream Cars API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080/Formation/DreamsCarsProject/BackEnd/public/api/",
            "description": "API Dream Cars"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get a lisr of category",
                "operationId": "category",
                "responses": {
                    "404": {
                        "description": "No found category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponseDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Return a list of category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriesChoiceDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/product/list": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Array of product",
                "operationId": "App\\Controller\\ProductController::productListAction",
                "responses": {
                    "404": {
                        "description": "Product no found in data base",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponseDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Return an array of product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/product/edit/{productId}": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update Product",
                "operationId": "update",
                "parameters": [
                    {
                        "parameter": "productId",
                        "name": "productId",
                        "in": "path",
                        "description": "Id for the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update Product",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductForm"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponseDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Form is invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponseDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found or Supplier not found or category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponseDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Return the product update",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/product/{productId}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get a product",
                "operationId": "product",
                "parameters": [
                    {
                        "parameter": "productId",
                        "name": "productId",
                        "in": "path",
                        "description": "Id for found a Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponseDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Return an array of product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/suppliers": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Get a list of choice supplier",
                "operationId": "supplier",
                "responses": {
                    "404": {
                        "description": "No found suppliers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponseDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Return a list of Supplier",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuppliersChoiceDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiErrorResponseDTO": {
                "title": "ApiErrorReponse Model",
                "description": "Api Error Response model",
                "properties": {
                    "statusCode": {
                        "description": "Status Error code",
                        "type": "integer"
                    },
                    "message": {
                        "description": "Error message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CategoriesChoiceDTO": {
                "title": "Category model",
                "description": "Category model",
                "properties": {
                    "id": {
                        "description": "Category Id",
                        "type": "string"
                    },
                    "name": {
                        "description": "Category Name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ProductDTO": {
                "title": "Product Model",
                "description": "Product Model",
                "properties": {
                    "id": {
                        "description": "Product id",
                        "type": "integer"
                    },
                    "product": {
                        "description": "Product name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Product's description",
                        "type": "string"
                    },
                    "price": {
                        "description": "Price of product",
                        "type": "integer"
                    },
                    "category": {
                        "description": "Product Category",
                        "type": "string"
                    },
                    "picture": {
                        "description": "Picture of product",
                        "type": "string"
                    },
                    "supplier": {
                        "description": "Supplier of product",
                        "type": "string"
                    },
                    "avaibility": {
                        "description": "Stock available for the product",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "SuppliersChoiceDTO": {
                "title": "Supplier Model",
                "description": "Supplier Model",
                "properties": {
                    "id": {
                        "description": "Product id",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Product Name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ProductForm": {
                "title": "Product Form Model",
                "description": "Product Form Model",
                "properties": {
                    "product": {
                        "description": "Product name",
                        "type": "string"
                    },
                    "price": {
                        "description": "Price of product",
                        "type": "integer"
                    },
                    "picture": {
                        "description": "Picture of product",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of product",
                        "type": "string"
                    },
                    "avaibility": {
                        "description": "Boolean of product",
                        "type": "boolean"
                    },
                    "category": {
                        "description": "Category of user",
                        "type": "object"
                    },
                    "supplier": {
                        "description": "Supplier of product",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}